requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# discord/src/bruh.py: 18
PyDictionary == 2.0.1

# discord/discord.py/discord/reader.py: 43,44
# discord/discord.py/discord/voice_client.py: 58
PyNaCl == 1.3.0

# discord/discord.py/discord/__main__.py: 33
# discord/discord.py/discord/client.py: 34
# discord/discord.py/discord/ext/tasks/__init__.py: 29
# discord/discord.py/discord/gateway.py: 39
# discord/discord.py/discord/http.py: 34
# discord/discord.py/discord/shard.py: 31
# discord/discord.py/discord/webhook.py: 33
aiohttp == 3.8.1

# discord/src/bruh.py: 8
# discord/tests/testapex.py: 1
apex_legends == 0.1.3

# discord/src/music.py: 12
async_timeout == 4.0.2

# discord/src/events.py: 6
# discord/src/music.py: 21
# discord/src/music_old.py: 9
# discord/tests/testbillboard.py: 1
# discord/tests/testwiki.py: 2
beautifulsoup4 == 4.5.3

# discord/discord.py/discord/ext/commands/bot.py: 36
# discord/discord.py/discord/ext/commands/context.py: 27,28
# discord/discord.py/discord/ext/commands/converter.py: 31
# discord/discord.py/discord/ext/commands/cooldowns.py: 27
# discord/discord.py/discord/ext/commands/core.py: 33
# discord/discord.py/discord/ext/commands/errors.py: 27
# discord/discord.py/discord/ext/commands/help.py: 32
# discord/discord.py/discord/ext/tasks/__init__.py: 30,36
# discord/discord.py/examples/basic_bot.py: 2
# discord/discord.py/examples/basic_voice.py: 6
# discord/discord.py/examples/custom_context.py: 4
# discord/src/audio/DiscordPCMStream.py: 3
# discord/src/audio/TranscriptionSink.py: 1,2
# discord/src/bot.py: 3,4
# discord/src/bruh.py: 1,2
# discord/src/card.py: 5,6
# discord/src/events.py: 1,3,9
# discord/src/music.py: 5,6,15
# discord/src/music_old.py: 2,7
# discord/src/speech.py: 2,3
# discord/tests/async_dont_work_in_jupyter.py: 5
# discord/tests/async_tests.py: 7
# discord/tests/better_audio_tests.py: 5
# discord/tests/old_bot.py: 1,2
### discord.py == 1.2.4 outdated
# git+https://github.com/sirmammingtonham/discord.py#egg=discord.py
discord.py == 2.0.1

# discord/discord.py/docs/extensions/details.py: 2,3,4
# discord/discord.py/docs/extensions/exception_hierarchy.py: 2,3,4
docutils == 0.16

# discord/hearthstone/Game.py: 2
# discord/hearthstone/gameUtils.py: 4,5,6,7
# discord/src/card.py: 7
### fireplace == 0.1.0 outdated
git+https://github.com/jleclanche/fireplace.git@58c5c60dc4502956d1a7a12c616c9b6201598edb#egg=fireplace

# discord/src/audio/TranscriptionSink.py: 18,19,20
google_api_python_client == 1.10.0

# discord/tests/cumber_tests.ipynb: 25
matplotlib == 3.3.0

# discord/src/events.py: 7
# discord/src/music.py: 20
# discord/src/music_old.py: 8
# discord/tests/testwiki.py: 3
mediawikiapi == 1.1.3

# discord/src/bruh.py: 10
# discord/src/verbosify.py: 3,4
# discord/tests/.ipynb_checkpoints/better_verbosify_tests-checkpoint.ipynb: 3
# discord/tests/better_verbosify_tests.ipynb: 3
# discord/tests/willc_tests.ipynb: 8,9,92
nltk == 3.5

# discord/hearthstone/gameUtils.py: 2
# discord/src/bruh.py: 12
# discord/tests/cumber_tests.ipynb: 2,34
numpy >= 1.15.4

# discord/src/audio/TranscriptionSink.py: 17
oauth2client == 4.1.3

# discord/src/bruh.py: 11
# discord/tests/cumber_tests.ipynb: 1,24,33
opencv-python >= 4.3.0.38

# discord/src/bruh.py: 7
# discord/tests/old_bot.py: 7
python_dateutil == 2.8.0

# discord/discord.py/discord/webhook.py: 263
# discord/src/bruh.py: 6
# discord/src/music.py: 18
# discord/src/music_old.py: 4
# discord/tests/testbillboard.py: 2
requests == 2.18.4

# discord/discord.py/discord/__main__.py: 32
# discord/discord.py/setup.py: 1
setuptools == 39.0.1

# discord/discord.py/docs/extensions/builder.py: 1,2
# discord/discord.py/docs/extensions/exception_hierarchy.py: 5
sphinx == 3.2.0

# discord/discord.py/examples/basic_voice.py: 4
# discord/src/music.py: 14
# discord/tests/old_bot.py: 3
# youtube_dl == 2019.11.5
wavelink>=1.2.4